build: refactor and add tests
