+++ package.json
@@ -2,14 +2,17 @@
   "name": "mjml-server",
   "private": true,
   "main": "lib/index.js",
-  "bin": {
-    "mjml-server": "bin/mjml-server"
+  "scripts": {
+    "test": "jest"
   },
   "dependencies": {
     "body-parser": "^1.20.2",
-    "commander": "^10.0.1",
     "express": "^4.18.2",
     "mjml": "^4.14.1",
     "newrelic": "^10.1.0"
+  },
+  "devDependencies": {
+    "jest": "^29.5.0",
+    "supertest": "^6.3.3"
   }
 }
